+++ hahomematic/platforms/hub/button.py
@@ -10,8 +10,7 @@ from typing import Final
 
 from hahomematic import central as hmcu
 from hahomematic.const import PROGRAM_ADDRESS, HmPlatform, HubData, ProgramData
-from hahomematic.platforms.decorators import state_property
-from hahomematic.platforms.entity import get_service_calls, service_call
+from hahomematic.platforms.decorators import get_service_calls, service, state_property
 from hahomematic.platforms.hub.entity import GenericHubEntity
 
 
@@ -64,7 +63,7 @@ class HmProgramButton(GenericHubEntity):
         if do_update:
             self.fire_entity_updated_callback()
 
-    @service_call()
+    @service()
     async def press(self) -> None:
         """Handle the button press."""
         await self.central.execute_program(pid=self.pid)