+++ changelog.md
@@ -1,6 +1,9 @@
 # Version 2024.9.13 (2024-09-29)
 
 - Make UPDATEABLE optional due to homegear support
+- Raise exception on set_value, put_paramset
+- Improve logging when raising exception
+- Log exception at the most outer service
 
 # Version 2024.9.12 (2024-09-26)